+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+1.3.1 - 31 DEC 2012
+-------------------
+
+* Promise has been updated to pass the promise/A+ specs according to
+https://github.com/promises-aplus/promises-tests
+* Updated StateMachine so new transitions can be added on the fly
+* Moved the CouchDB handler to CouchDB Emily Tools
+
 1.3.0 - 16 DEC 2012
 -------------------